requirements.txt
# Core AI and LLM
openai>=1.12.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Analytics
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web and API (for Render deployment)
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# UI/Frontend
gradio>=4.0.0
plotly>=5.0.0

# Database
sqlalchemy>=2.0.0
sqlite3worker>=1.1.7

# Social Media APIs
tweepy>=4.14.0
facebook-sdk>=3.1.0
instagrapi>=2.0.0

# Email
sendgrid>=6.10.0
jinja2>=3.1.2

# Scheduling
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
